revert changes to snapshots

+++ deeplabcut/pose_estimation_pytorch/runners/snapshots.py
@@ -113,26 +113,7 @@ class TorchSnapshotManager:
         ):
             current_best = self.best()
             self._best_metric = metrics[self._key]
-            
-            # Save the new best model
             save_path = self.snapshot_path(epoch, best=True)
-            parsed_state_dict = {
-                k: v 
-                for k, v in state_dict.items()
-                if self.save_optimizer_state or k != "optimizer"
-            }
-            torch.save(parsed_state_dict, save_path)
-
-            # Handle previous best model
-            if current_best is not None:
-                if current_best.epochs % self.save_epochs == 0:
-                    new_name = self.snapshot_path(epoch=current_best.epochs)
-                    current_best.path.rename(new_name)
-                else:
-                    current_best.path.unlink(missing_ok=False)
-        elif last or epoch % self.save_epochs == 0:
-            # Save regular snapshot if needed
-            save_path = self.snapshot_path(epoch=epoch)
             parsed_state_dict = {
                 k: v
                 for k, v in state_dict.items()
@@ -140,13 +121,32 @@ class TorchSnapshotManager:
             }
             torch.save(parsed_state_dict, save_path)
 
-        # Clean up old snapshots if needed
+            if current_best is not None:
+                # rename if the current best should have been saved, otherwise delete
+                if current_best.epochs % self.save_epochs == 0:
+                    new_name = self.snapshot_path(epoch=current_best.epochs)
+                    current_best.path.rename(new_name)
+                else:
+                    current_best.path.unlink(missing_ok=False)
+            return
+
+        if not (last or epoch % self.save_epochs == 0):
+            return
+
         existing_snapshots = [s for s in self.snapshots() if not s.best]
         if len(existing_snapshots) >= self.max_snapshots:
-            num_to_delete = len(existing_snapshots) - self.max_snapshots
+            num_to_delete = 1 + len(existing_snapshots) - self.max_snapshots
             to_delete = existing_snapshots[:num_to_delete]
             for snapshot in to_delete:
                 snapshot.path.unlink(missing_ok=False)
+
+        save_path = self.snapshot_path(epoch=epoch)
+        parsed_state_dict = {
+            k: v
+            for k, v in state_dict.items()
+            if self.save_optimizer_state or k != "optimizer"
+        }
+        torch.save(parsed_state_dict, save_path)
 
     def best(self) -> Snapshot | None:
         """Returns: the path to the best snapshot, if it exists"""
